refactor(search): fix artist filtering and clean up unused hooks
