+++ frontend/src/search/hooks/index.ts
@@ -1,6 +1,10 @@
 // Re-export the main hook
 export { useSearchState } from "./useSearchState";
 
+// Re-export navigation hooks
+export { useArtistNavigation } from "./useArtistNavigation";
+export { useSearchRedirect } from "./use-search-redirect";
+
 // Re-export modular components for testing/debugging if needed
 export { initialSearchState } from "./useSearchState/core/initialSearchState";
 export { searchStateReducer } from "./useSearchState/core/searchStateReducer";
