+++ frontend/src/search/hooks/use-search-redirect.ts
