+++ frontend/src/search/hooks/useArtistNavigation.ts
